refactor: KEEP-356 group sandbox files under lib/sandbox
- lib/sandbox-child-source.ts -> lib/sandbox/child-source.ts
- lib/sandbox-client.ts -> lib/sandbox/client.ts
