+++ plugins/code/steps/run-code.ts
@@ -7,8 +7,8 @@ import { withPluginMetrics } from "@/lib/metrics/instrumentation/plugin";
 import {
   SANDBOX_CHILD_SOURCE as CHILD_SOURCE,
   SANDBOX_RESULT_SENTINEL as RESULT_SENTINEL,
-} from "@/lib/sandbox-child-source";
-import { runRemote } from "@/lib/sandbox-client";
+} from "@/lib/sandbox/child-source";
+import { runRemote } from "@/lib/sandbox/client";
 import { type StepInput, withStepLogging } from "@/lib/steps/step-handler";
 
 type LogEntry = {
